requirements.txt
aiogram==2.11.2
APScheduler==3.6.3
pymongo==3.11.0
aioredis==1.3.1
python-dotenv==0.15.0
secrets==1.0.2
